+++ src/main/java/com/handongapp/cms/controller/v1/NodeController.java
@@ -1,6 +1,6 @@
-package com.handongapp.cms.controller;
+package com.handongapp.cms.controller.v1;
 
-import com.handongapp.cms.dto.NodeDto;
+import com.handongapp.cms.dto.v1.NodeDto;
 import com.handongapp.cms.service.NodeService;
 //import com.handongapp.cms.service.validator.CourseHierarchyValidator;
 import jakarta.validation.Valid;
