+++ src/main/java/com/handongapp/cms/service/NodeService.java
@@ -1,6 +1,6 @@
 package com.handongapp.cms.service;
 
-import com.handongapp.cms.dto.NodeDto;
+import com.handongapp.cms.dto.v1.NodeDto;
 
 import java.util.List;
